Be more strict for openldap_access resource title
The openldap_access resource allows a lot of variations in the title for
declaring a resource, making it possible to skip passing parameters such
as `what` and `suffix`.  This flexibility however can confuse Puppet
when it is prefetching resources, leading to catalog compilation
failures.

Impose a more strict format for resource titles, and validate it with
tighter custom types to raise a hopefully meaningful error instead of a
Ruby error because something borked bad.